+++ install_ydk.sh
@@ -109,14 +109,10 @@ function init_py_env {
   if [[ $ydk_lang == "py" || $ydk_lang == "all" ]]; then
     print_msg "Checking installation of python shared libraries"
     ver=$(python3 -c "import sys;print('{}.{}'.format(sys.version_info.major,sys.version_info.minor))")
-    if [[ $ver. > "3.8." ]]; then
-      print_msg "YDK for python currently is not supported with Python version $ver"
-      print_msg "Please downgrade your Python installation to 3.8 or 3.7"
-      exit 1
-    fi
+    ver_minor=$(python3 -c "import sys;print('{}'.format(sys.version_info.minor))")
     os_type=$(uname)
     if [[ ${os_type} == "Linux" ]]; then
-      if [[ $ver. > "3.7." ]]; then
+      if [ $ver_minor -gt 7 ]; then
         ext="$ver.so"
       else
         ext="$ver"m.so
@@ -366,6 +362,11 @@ if [ -z \$GOPATH ]; then
 fi
 export CXX=/usr/bin/c++
 export CC=/usr/bin/cc
+export CGO_ENABLED=1
+export CGO_LDFLAGS_ALLOW=\"-fprofile-arcs|-ftest-coverage|--coverage\"
+if [[ \$go_version > \"1.11.\" ]]; then
+    go env -w GO111MODULE=off
+fi
 " >> .env
   fi
   if [[ -n $CMAKE_LIBRARY_PATH ]]; then
@@ -502,7 +503,7 @@ print_msg "Running OS type: $os_type"
 print_msg "OS info: $os_info"
 if [[ ${os_type} == "Linux" ]]; then
   if [[ ${os_info} == *"Ubuntu"* ]]; then
-    if [[ ${os_info} != *"xenial"* && ${os_info} != *"bionic"* && ${os_info} != *"focal"* ]]; then
+    if [[ ${os_info} != *"xenial"* && ${os_info} != *"bionic"* && ${os_info} != *"focal"* && ${os_info} != *"jammy"* ]]; then
         print_msg "WARNING! Unsupported Ubuntu distribution found. Will try the best efforts."
     fi
   elif [[ ${os_info} == *"fedora"* ]]; then